requirements.txt
# Core data processing
polars==0.20.15
duckdb==0.10.2
pyarrow==15.0.2
pandas==2.2.1          # still needed for Great Expectations compatibility

# Validation
great-expectations==0.18.12

# Data generation (for test data)
faker==24.3.0
numpy==1.26.4

# Live streaming
kafka-python==2.0.2
redis==5.0.3

# API connectivity
httpx==0.27.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9

# Pydantic for typed outputs
pydantic==2.7.1

# Environment
python-dotenv==1.0.1

langgraph==0.1.19
langchain>=0.2.22,<0.3
langchain-core>=0.2.22,<0.3

groq==0.11.0


fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
plotly==5.22.0

setuptools
wheel



# uvicorn api.main:app --reload --port 8000